requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
APScheduler==3.10.4
python-dateutil==2.8.2
Werkzeug==2.3.7